// src/types/workout.ts
// Assuming this is in your types/workout.ts
export interface Workout {
  id: string;
  name: string;
  created: string;
  beschreibung: string;
  trainingsplanMontag: string;
  trainingsplanDienstag: string;
  trainingsplanMittwoch: string;
  trainingsplanDonnerstag: string;
  trainingsplanFreitag: string;
  trainingsplanSamstag: string;
  trainingsplanSonntag: string;
}